Cleanup Deep run setting is now ignored Removed click speed throttle feature Removed CPS counter
added heroConfig setting in ch_bot_settings (description in my config)

While a hero is being leveled, the script will try to buy upgrades and pick clickables every 20 seconds. For the last hero/round (item in heroConfig) it won't pick coins so we ensure we have one for ascension.

MonsterClicker click interval is set to 2 ms.