Added new Debug window UI (always enabled) The script now moves to the next hero as soon as it's available. After a configurable number of moves (heroCounter), it will spend a configurable amount of time (duration) on the *last* hero. When the time finishes, a run is considered finished.
Some error checking for reading the "buy upgrades button color" is implemented. The color is updated everytime we "scrollToBottom()". In the Debug window on the right side current color is being displayed - this might varry due to the way PixelGetColor is implemented, GPU, Windows version etc. As long as the colors are constant or change only between runs - the script will run fine.